Combine validity and form test suites
