--- src/checkbox.test.validity.ts
@@ -1,237 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-expressions */
-
-import { elementUpdated, expect, fixture, html } from '@open-wc/testing';
-import GlideCoreCheckbox from './checkbox.js';
-
-GlideCoreCheckbox.shadowRootOptions.mode = 'open';
-
-it('is valid if unchecked but not required', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label"></glide-core-checkbox>`,
-  );
-
-  expect(component.validity.valid).to.be.true;
-  expect(component.validity?.valueMissing).to.be.false;
-  expect(component.checkValidity()).to.be.true;
-  expect(component.reportValidity()).to.be.true;
-});
-
-it('is valid but not aria-invalid after being checked when unchecked and required', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label" required></glide-core-checkbox>`,
-  );
-
-  component.click();
-
-  expect(component.validity.valid).to.be.true;
-  expect(component.validity?.valueMissing).to.be.false;
-  expect(component.checkValidity()).to.be.true;
-  expect(component.reportValidity()).to.be.true;
-
-  expect(
-    component.shadowRoot?.querySelector('input')?.getAttribute('aria-invalid'),
-  ).to.equal('false');
-});
-
-it('is invalid but not aria-invalid if unchecked and required', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label" required></glide-core-checkbox>`,
-  );
-
-  expect(component.validity.valid).to.be.false;
-  expect(component.validity?.valueMissing).to.be.true;
-  expect(component.checkValidity()).to.be.false;
-  expect(component.reportValidity()).to.be.false;
-
-  expect(
-    component.shadowRoot?.querySelector('input')?.getAttribute('aria-invalid'),
-  ).to.equal('false');
-});
-
-it('is invalid but not aria-invalid after being unchecked when required', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox
-      label="Label"
-      checked
-      required
-    ></glide-core-checkbox>`,
-  );
-
-  component.click();
-
-  expect(component.validity.valid).to.be.false;
-  expect(component.validity?.valueMissing).to.be.true;
-  expect(component.checkValidity()).to.be.false;
-  expect(component.reportValidity()).to.be.false;
-
-  expect(
-    component.shadowRoot?.querySelector('input')?.getAttribute('aria-invalid'),
-  ).to.equal('false');
-});
-
-it('is both invalid and valid if unchecked and required but disabled', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox
-      label="Label"
-      disabled
-      required
-    ></glide-core-checkbox>`,
-  );
-
-  expect(component.validity.valid).to.be.false;
-  expect(component.validity?.valueMissing).to.be.true;
-  expect(component.checkValidity()).to.be.true;
-  expect(component.reportValidity()).to.be.true;
-});
-
-it('is valid when `value` is empty and `required` is updated to `false` programmatically', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label" required></glide-core-checkbox>`,
-  );
-
-  expect(component.validity?.valid).to.be.false;
-  expect(component.validity?.valueMissing).to.be.true;
-  expect(component.checkValidity()).to.be.false;
-  expect(component.reportValidity()).to.be.false;
-
-  await elementUpdated(component);
-
-  expect(
-    component.shadowRoot?.querySelector('input')?.getAttribute('aria-invalid'),
-  ).to.equal('true');
-
-  component.required = false;
-
-  await elementUpdated(component);
-
-  expect(component.validity?.valid).to.be.true;
-  expect(component.validity?.valueMissing).to.be.false;
-  expect(component.checkValidity()).to.be.true;
-  expect(component.reportValidity()).to.be.true;
-
-  expect(
-    component.shadowRoot?.querySelector('input')?.getAttribute('aria-invalid'),
-  ).to.equal('false');
-});
-
-it('sets the validity message with `setCustomValidity()`', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label"></glide-core-checkbox>`,
-  );
-
-  component.setCustomValidity('validity message');
-
-  expect(component.validity?.valid).to.be.false;
-  expect(component.validity?.customError).to.be.true;
-  expect(component.checkValidity()).to.be.false;
-
-  await elementUpdated(component);
-
-  // Like native, the validity message shouldn't display until `reportValidity()` is called.
-  expect(
-    component.shadowRoot?.querySelector('[data-test="validity-message"]')
-      ?.textContent,
-  ).to.be.undefined;
-
-  expect(component.reportValidity()).to.be.false;
-
-  await elementUpdated(component);
-
-  expect(
-    component.shadowRoot?.querySelector('input')?.getAttribute('aria-invalid'),
-  ).to.equal('true');
-
-  expect(
-    component.shadowRoot?.querySelector('[data-test="validity-message"]')
-      ?.textContent,
-  ).to.equal('validity message');
-});
-
-it('removes a validity message with an empty argument to `setCustomValidity()`', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label"></glide-core-checkbox>`,
-  );
-
-  component.setCustomValidity('validity message');
-  component.reportValidity();
-
-  await elementUpdated(component);
-
-  component.setCustomValidity('');
-
-  await elementUpdated(component);
-
-  expect(
-    component.shadowRoot?.querySelector('[data-test="validity-message"]')
-      ?.textContent,
-  ).to.be.undefined;
-});
-
-it('is invalid when `setValidity()` is called', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label"></glide-core-checkbox>`,
-  );
-
-  component.setValidity({ customError: true }, 'validity message');
-
-  expect(component.validity.valid).to.be.false;
-
-  await elementUpdated(component);
-
-  // Like native, the validity message shouldn't display until `reportValidity()` is called.
-  expect(
-    component.shadowRoot?.querySelector('[data-test="validity-message"]')
-      ?.textContent,
-  ).to.be.undefined;
-
-  expect(component.validity?.customError).to.be.true;
-
-  component.reportValidity();
-
-  await elementUpdated(component);
-
-  expect(
-    component.shadowRoot?.querySelector('input')?.getAttribute('aria-invalid'),
-  ).to.equal('true');
-
-  expect(
-    component.shadowRoot?.querySelector('[data-test="validity-message"]')
-      ?.textContent,
-  ).to.equal('validity message');
-});
-
-it('is valid when `setValidity()` is called', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label"></glide-core-checkbox>`,
-  );
-
-  component.setValidity({ customError: true }, 'validity message');
-
-  component.setValidity({});
-
-  await elementUpdated(component);
-
-  expect(component.validity.valid).to.be.true;
-  expect(component.validity.customError).to.be.false;
-
-  expect(component.reportValidity()).to.be.true;
-
-  await elementUpdated(component);
-
-  expect(
-    component.shadowRoot?.querySelector('[data-test="validity-message"]')
-      ?.textContent,
-  ).to.be.undefined;
-});
-
-it('retains existing validity state when `setCustomValidity()` is called', async () => {
-  const component = await fixture<GlideCoreCheckbox>(
-    html`<glide-core-checkbox label="Label" required></glide-core-checkbox>`,
-  );
-
-  component.setCustomValidity('validity message');
-
-  expect(component.validity?.valid).to.be.false;
-  expect(component.validity?.customError).to.be.true;
-  expect(component.validity?.valueMissing).to.be.true;
-});
